refactor to be similar to NTFSFindFast

+++ NtfsJournal/Support/LocaleFmt.cpp
@@ -2,9 +2,33 @@
 // Handle Locale (Language and Country) specific formatting.
 // This was derived from publicly available source code XFormatNumber 
 // http://www.codeproject.com/script/Articles/ViewDownloads.aspx?aid=2492
+// License:  The Code Project Open License (CPOL) 
+//           http://www.codeproject.com/info/cpol10.aspx
 //
 // Author:  Dennis Lang   Apr-2011
 // https://lanenlabs.com
+//
+// ----- License ----
+//
+// Copyright (c) 2014 Dennis Lang
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
+// of the Software, and to permit persons to whom the Software is furnished to do
+// so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+// FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+// COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+// IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+// CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+//
 // ------------------------------------------------------------------------------------------------
 
 #include "LocaleFmt.h"
@@ -41,8 +65,8 @@ wchar_t* LocaleFmt::snprintf(wchar_t* str, unsigned maxChar, wchar_t* fmt, ...)
 
     // Get number of digits right of decimal point.
     const wchar_t* pDec = wcsstr(str, nf.lpDecimalSep);
-    unsigned decimalPt = (pDec == NULL) ? 0 : (unsigned)wcslen(pDec) - (unsigned)wcslen(nf.lpDecimalSep);
-    nf.NumDigits = decimalPt;
+    size_t decimalPt = (pDec == NULL) ? 0 : wcslen(pDec) - wcslen(nf.lpDecimalSep);
+    nf.NumDigits = (UINT)decimalPt;
 
     // Copy raw string into temporary buffer.
     std::wstring rawStr = str;
